bot_requirements.txt
ccxt>=4.0.0
email-validator>=2.0.0
eventlet>=0.33.0
flask>=2.0.0
flask-socketio>=5.3.0
flask-sqlalchemy>=3.0.0
gunicorn>=21.0.0
numpy>=1.22.0
pandas>=1.4.0
pandas-ta>=0.3.0
scikit-learn>=1.0.0
sqlalchemy>=2.0.0
pyinstaller>=6.0.0
joblib>=1.1.0
psycopg2-binary>=2.9.0
python-socketio>=5.8.0
python-engineio>=4.5.0
websocket-client>=1.5.0
dnspython>=2.3.0
bidict>=0.22.1
simple-websocket>=0.10.0
python-dateutil>=2.8.2
gevent>=22.10.2
h11>=0.14.0
pyzmq>=25.1.0